+++ dev/webdock/basebox/nodejs-foundation/hooks/test.sh
@@ -1,11 +1,10 @@
 #!/usr/bin/env sh
 set -e
+. "$(dirname "$0")/_variables.sh"
 
-echo '[Hook] Running installation tests...'
-docker-compose run --rm this bash -c '
+echo 'Running installation tests...'
+docker run --rm "${IMAGE}:${VERSION}" bash -c '
   node --version \
   && npm --version \
   && git --version
 '
-
-echo '[Hook] Done.'