+++ dev/webdock/basebox/nodejs-web-server-foundation/hooks/_variables.sh
@@ -1,0 +1,2 @@
+IMAGE=127.0.0.1:5000/webdock/nodejs-web-server-foundation
+VERSION="${VERSION:-unreleased}"