+++ dev/webdock/basebox/php-web-server-foundation/bin/entrypoint.bash
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 set -e
 
-substitute_environment_variables()
+inject_environment_variables()
 {
   envsubst $(printenv | cut -f1 -d'=' | sed 's/.*/\\\${&}/' | tr '\n' ',')
 }
 
-substitute_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
+inject_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
 
 if [ "$XDEBUG_ENABLE" == 'on' ]
 then