+++ dev/webdock/basebox/php-web-server-foundation/hooks/test.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env sh
 set -e
+. "$(dirname "$0")/_variables.sh"
 
-echo '[Hook] Running installation tests...'
-docker-compose run --rm this bash -c '
+echo 'Running installation tests...'
+docker run --rm "${IMAGE}:${VERSION}" bash -c '
   php --version \
   && php-fpm --version \
   && pear version \
@@ -14,24 +15,22 @@ docker-compose run --rm this bash -c '
   && curl --version \
   && git --version
 '
-docker-compose run --rm this bash -c '
+docker run --rm "${IMAGE}:${VERSION}" bash -c '
   if php -r "exit(extension_loaded(\"xdebug\") ? 1 : 0);"
   then
-    echo "[OK] Xdebug is disabled by default."
+    echo "OK: Xdebug is disabled by default."
   else
-    echo "[ERROR] Xdebug is enabled by default."
+    echo "ERROR: Xdebug is enabled by default."
     exit 1
   fi
 '
-docker-compose run --rm this bash -c '
+docker run --rm "${IMAGE}:${VERSION}" bash -c '
   docker-php-ext-enable xdebug
   if php -r "exit(extension_loaded(\"xdebug\") ? 1 : 0);"
   then
-    echo "[ERROR] Xdebug can not be enabled."
+    echo "ERROR: Xdebug can not be enabled."
     exit 1
   else
-    echo "[OK] Xdebug can be enabled."
+    echo "OK: Xdebug can be enabled."
   fi
 '
-
-echo '[Hook] Done.'