+++ dev/webdock/basebox/web-assets-builder/hooks/_variables.sh
@@ -1,0 +1,2 @@
+IMAGE=127.0.0.1:5000/webdock/web-assets-builder
+VERSION="${VERSION:-unreleased}"