+++ dev/webdock/basebox/web-assets-builder/hooks/test.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env sh
 set -e
+. "$(dirname "$0")/_variables.sh"
 
-echo '[Hook] Running installation tests...'
-docker-compose run --rm this bash -c "
+echo 'Running installation tests...'
+docker run --rm "${IMAGE}:${VERSION}" bash -c '
   node --version \
   && npm --version \
   && git --version \
@@ -10,5 +11,4 @@ docker-compose run --rm this bash -c "
   && tsc --version \
   && webpack --version \
   && webpack-cli --version
-"
-echo '[Hook] Done.'
+'