+++ dev/webdock/basebox/web-client-foundation/bin/entrypoint.bash
@@ -1,17 +1,17 @@
 #!/usr/bin/env bash
 set -e
 
-substitute_environment_variables()
+inject_environment_variables()
 {
   envsubst $(printenv | cut -f1 -d'=' | sed 's/.*/\\\${&}/' | tr '\n' ',')
 }
 
-substitute_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
+inject_environment_variables < /etc/nginx/nginx.template.conf > /etc/nginx/nginx.conf
 
-if [ -d /app/templates ]
+if [ -d /app/build-templates ]
 then
-  for template in $(ls /app/templates); do
-    substitute_environment_variables < /app/templates/$template > /app/build/$template
+  for file in $(cd /app/build-templates; find . -maxdepth 1 -type f); do
+    inject_environment_variables < /app/build-templates/$file > /app/build/$file
   done
 fi