+++ dev/webdock/domain/nodejs-web-server-app/hooks/_variables.sh
@@ -1,0 +1,2 @@
+IMAGE=127.0.0.1:5000/webdock/nodejs-web-server-app
+VERSION="${VERSION:-unreleased}"