+++ dev/webdock/domain/nodejs-web-server-compose/hooks/deploy.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env sh
 set -e
 
-echo '[Hook] Creating containers...'
+echo 'Creating containers...'
 docker-compose up --detach --remove-orphans
-
-echo '[Hook] Done.'