+++ dev/webdock/domain/nodejs-web-server-compose/hooks/test.sh
@@ -1,8 +1,6 @@
 #!/usr/bin/env sh
 set -e
 
-echo '[Hook] Running HTTP API tests...'
+echo 'Running HTTP API tests...'
 docker-compose exec http_tester \
   npm run target-http-api-tests
-
-echo '[Hook] Done.'