+++ dev/webdock/domain/nodejs-web-server-tester/hooks/build.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env sh
 set -e
+. "$(dirname "$0")/_variables.sh"
 
-echo '[Hook] Building Docker image...'
-docker-compose build
-
-echo '[Hook] Done.'
+echo 'Building Docker image...'
+docker build --tag="${IMAGE}:${VERSION}" .