+++ dev/webdock/domain/php-web-server-tester/hooks/_variables.sh
@@ -1,0 +1,2 @@
+IMAGE=127.0.0.1:5000/webdock/php-web-server-tester
+VERSION="${VERSION:-unreleased}"