+++ dev/webdock/domain/php-web-server-tester/hooks/test.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env sh
 set -e
+. "$(dirname "$0")/_variables.sh"
 
-echo '[Hook] Running unit tests...'
-docker-compose run --rm this bash -c '
+echo 'Running unit tests...'
+docker run \
+  --rm \
+  --mount="type=volume,volume-driver=local,src=global_composer_cache,dst=/root/.composer/cache" \
+"${IMAGE}:${VERSION}" bash -c '
   composer install --quiet \
   && composer run-script unit-tests
 '
-
-echo '[Hook] Done.'