+++ dev/webdock/domain/web-client-compose/hooks/test.sh
@@ -1,8 +1,6 @@
 #!/usr/bin/env sh
 set -e
 
-echo '[Hook] Running GUI tests...'
+echo 'Running GUI tests...'
 docker-compose exec web_tester \
   npm run target-gui-tests
-
-echo '[Hook] Done.'