+++ dev/webdock/sandbox/backing-services-compose/hooks/test.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env sh
 set -e
 
-echo '[Hook] No tests specified.'
+echo 'No tests specified.'