+++ dev/webdock/domain/php-web-server-compose/hooks/test.bash
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-echo 'Running HTTP API tests...'
-docker-compose exec http_tester \
-  composer run-script target-http-api-tests
+echo 'Validating Docker Compose manifest...'
+docker-compose config --quiet