+++ dev/webdock/domain/web-client-compose/helpers/run-target-tests.bash
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+
+docker-compose exec web_tester \
+  npm run target-gui-tests