+++ dev/webdock/domain/web-client-compose/hooks/test.bash
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-echo 'Running GUI tests...'
-docker-compose exec web_tester \
-  npm run target-gui-tests
+echo 'Validating Docker Compose manifest...'
+docker-compose config --quiet