+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Run browser tests with WebDriverIO.
 - Replace Selenium Grid with own standalone Firefox container.
+- Control PHP, FPM and Nginx logging level with environment variables.
 
 ## [0.13.0] - 2019-04-02